requirements.txt
asgiref==3.11.0
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
